pyproject.toml
[project]
name = "antibody-abtigen"
version = "0.1.0"
description = "Cross-species antibody-antigen structure dataset builder"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "a-green-hand-jack"}
]
keywords = ["antibody", "antigen", "structure", "bioinformatics", "PDB", "SAbDab"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "biopython>=1.80",
    "click>=8.0",
    "pandas>=2.0",
    "requests>=2.28",
    "tqdm>=4.60",
    "gemmi>=0.6.0",
    "pyyaml>=6.0",
    "nvitop>=1.6.0",
    "torch>=2.0.0",
    "fair-esm>=2.0.0",
    "h5py>=3.8.0",
    "networkx>=3.0",
    "faiss-gpu>=1.7.2",
    "biotite>=0.37.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
# PyMOL is optional and only available on certain platforms (macOS, Windows, some Linux)
# On unsupported platforms, Biopython alignment will be used automatically
pymol = [
    "pymol-open-source>=3.0; sys_platform == 'darwin' or sys_platform == 'win32'"
]
dev = ["ruff", "pytest"]

[project.scripts]
antibody-abtigen = "antibody_abtigen.cli:main"

[project.urls]
Homepage = "https://github.com/a-green-hand-jack/antibody-abtigen"
Repository = "https://github.com/a-green-hand-jack/antibody-abtigen"
Issues = "https://github.com/a-green-hand-jack/antibody-abtigen/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["antibody_abtigen"]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
